requirements.txt
fastapi
sqlalchemy
uvicorn
python-decouple
psycopg2-binary
asyncpg
pydantic
python-jose
bcrypt